further refactor and add scm ip restrictions to function apps
