+++ azurerm/internal/services/web/resource_arm_function_app.go
@@ -1,7 +1,6 @@
 package web
 
 import (
-	"context"
 	"fmt"
 	"log"
 	"strings"
@@ -336,7 +335,7 @@ func resourceArmFunctionAppCreate(d *schema.ResourceData, meta interface{}) erro
 		sourceControlProperties := expandAppServiceSiteSourceControl(d)
 		sourceControl := &web.SiteSourceControl{}
 		sourceControl.SiteSourceControlProperties = sourceControlProperties
-		// TODO - Do we need to lock the app for updates?
+		// TODO - Do we need to lock the function app for updates?
 		_, err := client.CreateOrUpdateSourceControl(ctx, resourceGroup, name, *sourceControl)
 		if err != nil {
 			return fmt.Errorf("failed to create App Service Source Control for %q (Resource Group %q): %+v", name, resourceGroup, err)
@@ -691,355 +690,4 @@ func resourceArmFunctionAppDelete(d *schema.ResourceData, meta interface{}) erro
 	}
 
 	return nil
-}
-
-func getBasicFunctionAppAppSettings(d *schema.ResourceData, appServiceTier, endpointSuffix string) ([]web.NameValuePair, error) {
-	// TODO: This is a workaround since there are no public Functions API
-	// You may track the API request here: https://github.com/Azure/azure-rest-api-specs/issues/3750
-	dashboardPropName := "AzureWebJobsDashboard"
-	storagePropName := "AzureWebJobsStorage"
-	functionVersionPropName := "FUNCTIONS_EXTENSION_VERSION"
-
-	contentSharePropName := "WEBSITE_CONTENTSHARE"
-	contentFileConnStringPropName := "WEBSITE_CONTENTAZUREFILECONNECTIONSTRING"
-
-	// TODO 3.0 - remove this logic for determining which storage account connection string to use
-	storageConnection := ""
-	if v, ok := d.GetOk("storage_connection_string"); ok {
-		storageConnection = v.(string)
-	}
-
-	storageAccount := ""
-	if v, ok := d.GetOk("storage_account_name"); ok {
-		storageAccount = v.(string)
-	}
-
-	connectionString := ""
-	if v, ok := d.GetOk("storage_account_access_key"); ok {
-		connectionString = v.(string)
-	}
-
-	if storageConnection == "" && storageAccount == "" && connectionString == "" {
-		return nil, fmt.Errorf("one of `storage_connection_string` or `storage_account_name` and `storage_account_access_key` must be specified")
-	}
-
-	if (storageAccount == "" && connectionString != "") || (storageAccount != "" && connectionString == "") {
-		return nil, fmt.Errorf("both `storage_account_name` and `storage_account_access_key` must be specified")
-	}
-
-	if connectionString != "" && storageAccount != "" {
-		storageConnection = fmt.Sprintf("DefaultEndpointsProtocol=https;AccountName=%s;AccountKey=%s;EndpointSuffix=%s", storageAccount, connectionString, endpointSuffix)
-	}
-
-	functionVersion := d.Get("version").(string)
-	contentShare := strings.ToLower(d.Get("name").(string)) + "-content"
-
-	basicSettings := []web.NameValuePair{
-		{Name: &storagePropName, Value: &storageConnection},
-		{Name: &functionVersionPropName, Value: &functionVersion},
-	}
-
-	if d.Get("enable_builtin_logging").(bool) {
-		basicSettings = append(basicSettings, web.NameValuePair{
-			Name:  &dashboardPropName,
-			Value: &storageConnection,
-		})
-	}
-
-	consumptionSettings := []web.NameValuePair{
-		{Name: &contentSharePropName, Value: &contentShare},
-		{Name: &contentFileConnStringPropName, Value: &storageConnection},
-	}
-
-	// On consumption and premium plans include WEBSITE_CONTENT components, unless it's a Linux consumption plan
-	// (see https://github.com/Azure/azure-functions-python-worker/issues/598)
-	if (strings.EqualFold(appServiceTier, "dynamic") || strings.EqualFold(appServiceTier, "elasticpremium")) &&
-		!strings.EqualFold(d.Get("os_type").(string), "linux") {
-		return append(basicSettings, consumptionSettings...), nil
-	}
-
-	return basicSettings, nil
-}
-
-func getFunctionAppServiceTier(ctx context.Context, appServicePlanId string, meta interface{}) (string, error) {
-	id, err := ParseAppServicePlanID(appServicePlanId)
-	if err != nil {
-		return "", fmt.Errorf("[ERROR] Unable to parse App Service Plan ID %q: %+v", appServicePlanId, err)
-	}
-
-	log.Printf("[DEBUG] Retrieving App Service Plan %q (Resource Group %q)", id.Name, id.ResourceGroup)
-
-	appServicePlansClient := meta.(*clients.Client).Web.AppServicePlansClient
-	appServicePlan, err := appServicePlansClient.Get(ctx, id.ResourceGroup, id.Name)
-	if err != nil {
-		return "", fmt.Errorf("[ERROR] Could not retrieve App Service Plan ID %q: %+v", appServicePlanId, err)
-	}
-
-	if sku := appServicePlan.Sku; sku != nil {
-		if tier := sku.Tier; tier != nil {
-			return *tier, nil
-		}
-	}
-	return "", fmt.Errorf("No `sku` block was returned for App Service Plan ID %q", appServicePlanId)
-}
-
-func expandFunctionAppAppSettings(d *schema.ResourceData, appServiceTier, endpointSuffix string) (map[string]*string, error) {
-	output := expandAppServiceAppSettings(d)
-
-	basicAppSettings, err := getBasicFunctionAppAppSettings(d, appServiceTier, endpointSuffix)
-	if err != nil {
-		return nil, err
-	}
-	for _, p := range basicAppSettings {
-		output[*p.Name] = p.Value
-	}
-
-	return output, nil
-}
-
-func expandFunctionAppSiteConfig(d *schema.ResourceData) (web.SiteConfig, error) {
-	configs := d.Get("site_config").([]interface{})
-	siteConfig := web.SiteConfig{}
-
-	if len(configs) == 0 {
-		return siteConfig, nil
-	}
-
-	config := configs[0].(map[string]interface{})
-
-	if v, ok := config["always_on"]; ok {
-		siteConfig.AlwaysOn = utils.Bool(v.(bool))
-	}
-
-	if v, ok := config["use_32_bit_worker_process"]; ok {
-		siteConfig.Use32BitWorkerProcess = utils.Bool(v.(bool))
-	}
-
-	if v, ok := config["websockets_enabled"]; ok {
-		siteConfig.WebSocketsEnabled = utils.Bool(v.(bool))
-	}
-
-	if v, ok := config["linux_fx_version"]; ok {
-		siteConfig.LinuxFxVersion = utils.String(v.(string))
-	}
-
-	if v, ok := config["cors"]; ok {
-		corsSettings := v.(interface{})
-		expand := azure.ExpandWebCorsSettings(corsSettings)
-		siteConfig.Cors = &expand
-	}
-
-	if v, ok := config["http2_enabled"]; ok {
-		siteConfig.HTTP20Enabled = utils.Bool(v.(bool))
-	}
-
-	if v, ok := config["ip_restriction"]; ok {
-		ipSecurityRestrictions := v.(interface{})
-		restrictions, err := expandAppServiceIpRestriction(ipSecurityRestrictions)
-		if err != nil {
-			return siteConfig, err
-		}
-		siteConfig.IPSecurityRestrictions = &restrictions
-	}
-
-	if v, ok := config["min_tls_version"]; ok {
-		siteConfig.MinTLSVersion = web.SupportedTLSVersions(v.(string))
-	}
-
-	if v, ok := config["ftps_state"]; ok {
-		siteConfig.FtpsState = web.FtpsState(v.(string))
-	}
-
-	if v, ok := config["pre_warmed_instance_count"]; ok {
-		siteConfig.PreWarmedInstanceCount = utils.Int32(int32(v.(int)))
-	}
-
-	if v, ok := config["scm_type"]; ok {
-		siteConfig.ScmType = web.ScmType(v.(string))
-	}
-
-	return siteConfig, nil
-}
-
-func flattenFunctionAppSiteConfig(input *web.SiteConfig) []interface{} {
-	results := make([]interface{}, 0)
-	result := make(map[string]interface{})
-
-	if input == nil {
-		log.Printf("[DEBUG] SiteConfig is nil")
-		return results
-	}
-
-	if input.AlwaysOn != nil {
-		result["always_on"] = *input.AlwaysOn
-	}
-
-	if input.Use32BitWorkerProcess != nil {
-		result["use_32_bit_worker_process"] = *input.Use32BitWorkerProcess
-	}
-
-	if input.WebSocketsEnabled != nil {
-		result["websockets_enabled"] = *input.WebSocketsEnabled
-	}
-
-	if input.LinuxFxVersion != nil {
-		result["linux_fx_version"] = *input.LinuxFxVersion
-	}
-
-	if input.HTTP20Enabled != nil {
-		result["http2_enabled"] = *input.HTTP20Enabled
-	}
-
-	if input.PreWarmedInstanceCount != nil {
-		result["pre_warmed_instance_count"] = *input.PreWarmedInstanceCount
-	}
-
-	result["ip_restriction"] = flattenFunctionAppIpRestriction(input.IPSecurityRestrictions)
-
-	result["min_tls_version"] = string(input.MinTLSVersion)
-	result["ftps_state"] = string(input.FtpsState)
-
-	result["cors"] = azure.FlattenWebCorsSettings(input.Cors)
-
-	results = append(results, result)
-	return results
-}
-
-func expandFunctionAppConnectionStrings(d *schema.ResourceData) map[string]*web.ConnStringValueTypePair {
-	input := d.Get("connection_string").(*schema.Set).List()
-	output := make(map[string]*web.ConnStringValueTypePair, len(input))
-
-	for _, v := range input {
-		vals := v.(map[string]interface{})
-
-		csName := vals["name"].(string)
-		csType := vals["type"].(string)
-		csValue := vals["value"].(string)
-
-		output[csName] = &web.ConnStringValueTypePair{
-			Value: utils.String(csValue),
-			Type:  web.ConnectionStringType(csType),
-		}
-	}
-
-	return output
-}
-
-//func expandFunctionAppIpRestriction(input interface{}) ([]web.IPSecurityRestriction, error) {
-//	restrictions := make([]web.IPSecurityRestriction, 0)
-//
-//	for i, r := range input.([]interface{}) {
-//		if r == nil {
-//			continue
-//		}
-//
-//		restriction := r.(map[string]interface{})
-//
-//		ipAddress := restriction["ip_address"].(string)
-//		vNetSubnetID := restriction["subnet_id"].(string)
-//		name := restriction["name"].(string)
-//		priority := restriction["priority"].(int)
-//		action := restriction["action"].(string)
-//
-//		if vNetSubnetID != "" && ipAddress != "" {
-//			return nil, fmt.Errorf(fmt.Sprintf("only one of `ip_address` or `subnet_id` can set for `site_config.0.ip_restriction.%d`", i))
-//		}
-//
-//		if vNetSubnetID == "" && ipAddress == "" {
-//			return nil, fmt.Errorf(fmt.Sprintf("one of `ip_address` or `subnet_id` must be set for `site_config.0.ip_restriction.%d`", i))
-//		}
-//
-//		ipSecurityRestriction := web.IPSecurityRestriction{}
-//		if ipAddress == "Any" {
-//			continue
-//		}
-//
-//		if ipAddress != "" {
-//			ipSecurityRestriction.IPAddress = &ipAddress
-//		}
-//
-//		if vNetSubnetID != "" {
-//			ipSecurityRestriction.VnetSubnetResourceID = &vNetSubnetID
-//		}
-//
-//		if name != "" {
-//			ipSecurityRestriction.Name = &name
-//		}
-//
-//		if priority != 0 {
-//			ipSecurityRestriction.Priority = utils.Int32(int32(priority))
-//		}
-//
-//		if action != "" {
-//			ipSecurityRestriction.Action = &action
-//		}
-//
-//		restrictions = append(restrictions, ipSecurityRestriction)
-//	}
-//
-//	return restrictions, nil
-//}
-
-func flattenFunctionAppConnectionStrings(input map[string]*web.ConnStringValueTypePair) interface{} {
-	results := make([]interface{}, 0)
-
-	for k, v := range input {
-		result := make(map[string]interface{})
-		result["name"] = k
-		result["type"] = string(v.Type)
-		result["value"] = *v.Value
-		results = append(results, result)
-	}
-
-	return results
-}
-
-func flattenFunctionAppSiteCredential(input *web.UserProperties) []interface{} {
-	results := make([]interface{}, 0)
-	result := make(map[string]interface{})
-
-	if input == nil {
-		log.Printf("[DEBUG] UserProperties is nil")
-		return results
-	}
-
-	if input.PublishingUserName != nil {
-		result["username"] = *input.PublishingUserName
-	}
-
-	if input.PublishingPassword != nil {
-		result["password"] = *input.PublishingPassword
-	}
-
-	return append(results, result)
-}
-
-func flattenFunctionAppIpRestriction(input *[]web.IPSecurityRestriction) []interface{} {
-	restrictions := make([]interface{}, 0)
-
-	if input == nil {
-		return restrictions
-	}
-
-	for _, v := range *input {
-		ipAddress := ""
-		if v.IPAddress != nil {
-			ipAddress = *v.IPAddress
-			if ipAddress == "Any" {
-				continue
-			}
-		}
-
-		subnetId := ""
-		if v.VnetSubnetResourceID != nil {
-			subnetId = *v.VnetSubnetResourceID
-		}
-
-		restrictions = append(restrictions, map[string]interface{}{
-			"ip_address": ipAddress,
-			"subnet_id":  subnetId,
-		})
-	}
-
-	return restrictions
 }
